refactor gnosis package
